+++ cmds/account.go
@@ -1,0 +1,77 @@
+package cmds
+
+import (
+	"encoding/json"
+	"fmt"
+	"github.com/bleuhold/bh/filesys"
+	"github.com/dottics/cli"
+	"github.com/google/uuid"
+	"log"
+)
+
+var ACCOUNT *cli.Command
+var accountsFilename = "accounts.json"
+
+type Account struct {
+	UUID         uuid.UUID
+	Number       string
+	AccountType  string
+	ProviderName string
+}
+
+// NewAccount creates a new account, with a default UUID.
+func NewAccount(number, accountType, providerName string) *Account {
+	a := &Account{
+		UUID:         uuid.New(),
+		Number:       number,
+		AccountType:  accountType,
+		ProviderName: providerName,
+	}
+	return a
+}
+
+type Accounts []Account
+
+// LoadAccounts loads the accounts' information from the file system.
+func LoadAccounts() *Accounts {
+	xa := &Accounts{}
+	filesys.LoadInterface(accountsFilename, xa)
+	return xa
+}
+
+// Save writes the accounts' data to the file system.
+func (xa *Accounts) Save() {
+	xb, err := json.Marshal(xa)
+	if err != nil {
+		log.Fatalf("Unable to marshal accounts data: %v", err)
+	}
+	filesys.WriteFile(accountsFilename, xb)
+}
+
+func (xa *Accounts) String() string {
+	s := ""
+	for _, a := range *xa {
+		s += fmt.Sprintf("%36s %16s %16s %16s\n", a.UUID, a.Number, a.ProviderName, a.AccountType)
+	}
+	return s
+}
+
+/*
+	ACCOUNT
+*/
+
+func AccountExecute(cmd *cli.Command) error {
+	switch {
+	case B1:
+		ListAccounts()
+		return nil
+	default:
+		cmd.PrintHelp()
+	}
+	return nil
+}
+
+func ListAccounts() {
+	xa := LoadAccounts()
+	fmt.Print(xa.String())
+}